Bring 406 upstream commits onto the SDL3 migration branch. 29 conflicted
files resolved across 12 rename/move conflicts (DU) and 17 content
conflicts (UU).

Key upstream fixes ported:
- Walk-jitter (PR #324): + FPS_ANIMATION_FACTOR -> * FPS_ANIMATION_FACTOR
- Helper-attack/heal: IsMonster() filtering, party-index increment fix,
  pTarget->TargetX/TargetX typo, self-position-skill handling for
  Nova/Hellfire/Inferno
- Right-click drop (PR #311): IInventoryActionContext interface +
  CNewUIInventoryActionController; ~1500 lines of static helpers moved
  off CNewUIMyInventory into the controller
- Phantom items (PR #317): inventory resync on pickup
- Tooltips FPS (PR #321): GroundItemLabelDescriptor + budget
- Exp-bar overflow (PR #316): saturating arithmetic, __int64 widening
- Divine stick (PR #306): IsDivineArchangelWeaponItem/Model helpers
- defenseSuccessRate calc fix in CNewUICharacterInfoWindow
- Lucky-set/restricted-item constants extracted to mu_define.h
- DetermineMonsterObjectKind extraction in ZzzCharacter
- VectorScale double-scaling removal in ZzzEffectJoint

SDL migration changes preserved verbatim where they don't conflict;
SDL3-only paths (renderer QueueTextureUpdate, packet queue) chosen over
the OpenGL alternatives main added.

Note: 4 files added by main (IInventoryActionContext.h, InventoryUtils.h,
NewUIInventoryActionController.{h,cpp}) moved from src/source/ root into
src/source/UI/Windows/Inventory/ to match the new module layout. CMake
globs already include this directory.

+++ src/source/UI/Windows/HUD/NewUIMainFrameWindow.h
@@ -219,8 +219,8 @@ public:
     void SetPreExp_Wide(__int64 dwPreExp);
     void SetGetExp_Wide(__int64 dwGetExp);
 
-    void SetPreExp(DWORD dwPreExp);
-    void SetGetExp(DWORD dwGetExp);
+    void SetPreExp(__int64 dwPreExp);
+    void SetGetExp(__int64 dwGetExp);
 
     // buttons
     void SetBtnState(int iBtnType, bool bStateDown);
@@ -259,8 +259,8 @@ private:
     bool m_bExpEffect;
     DWORD m_dwExpEffectTime;
 
-    DWORD m_dwPreExp;
-    DWORD m_dwGetExp;
+    __int64 m_dwPreExp;
+    __int64 m_dwGetExp;
 
 #ifdef PBG_ADD_INGAMESHOP_UI_MAINFRAME
     CNewUIButton m_BtnCShop;
